Properly unmount React containers
