--- web_src/sidepanel/initSidepanel.jsx
@@ -1,36 +1,0 @@
-/* global L */
-
-import { createRoot } from 'react-dom/client';
-
-export let sidepanelRoot = null;
-
-/* Initializes the DOM to support the sidepanel. */
-export function initSidepanelDom() {
-  const sidepanelContainer = document.getElementById('sidepanel');
-  // sidepanelContainer.replaceChildren(); // Clear all child elements
-
-  // Adds stopPropagation to the element's 'wheel' events (plus browser variants).
-  L.DomEvent.disableScrollPropagation(sidepanelContainer);
-  // Adds stopPropagation to the element's 'click', 'dblclick', 'contextmenu', 'mousedown'
-  // and 'touchstart' events (plus browser variants).
-  L.DomEvent.disableClickPropagation(sidepanelContainer);
-
-  sidepanelRoot = createRoot(sidepanelContainer);
-}
-
-/* Sets up CSS on the Leaflet map so that any left-side controls are pushed to the right while the sidebar
- * is open (so that the controls are not hidden by the sidebar). */
-export const setLeafletMapPushCss = (sidebarIsOpen) => {
-  let leaflet_map = document.querySelector('.leaflet-control-container');
-  if (!leaflet_map) {
-    return;
-  }
-  leaflet_map.classList.add('leaflet-anim-control-container');
-  if (sidebarIsOpen) {
-    leaflet_map.classList.add('left-opened');
-    leaflet_map.classList.remove('left-closed');
-  } else {
-    leaflet_map.classList.add('left-closed');
-    leaflet_map.classList.remove('left-opened');
-  }
-};
